# 03-API-SPEC.yaml
openapi: 3.0.3
info:
  title: Saudi GRC Platform API
  description: |
    Multi-tenant GRC Compliance Platform API built with ABP.io Open Source.
    AI Agent: Use this specification to implement API controllers.
  version: 1.0.0
  contact:
    name: GRC Platform
    email: api@grc-platform.sa
servers:
  - url: https://{tenant}.grc-platform.sa/api
    description: Production
    variables:
      tenant:
        default: demo
        description: Tenant subdomain
  - url: http://localhost:5000/api
    description: Development
security:
  - bearerAuth: []
  - oauth2: []
...
paths:
...
  /grc/subscriptions/current:
    get:
      tags: [Subscriptions]
      operationId: getCurrentSubscription
      summary: Get tenant's current subscription
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionDetailDto'
        '404':
          $ref: '#/components/responses/NotFound'
  /grc/subscriptions/subscribe:
    post:
      tags: [Subscriptions]
      operationId: subscribeToProduct
      summary: Subscribe to a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscribeInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantSubscriptionDto'
  /grc/subscriptions/{id}/cancel:
    post:
      tags: [Subscriptions]
      operationId: cancelSubscription
      summary: Cancel subscription
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelSubscriptionInput'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantSubscriptionDto'
  /grc/subscriptions/upgrade:
    post:
      tags: [Subscriptions]
      operationId: upgradeSubscription
      summary: Upgrade subscription to a different product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpgradeSubscriptionInput'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantSubscriptionDto'
  /grc/subscriptions/quota-usage:
    get:
      tags: [Subscriptions]
      operationId: getQuotaUsage
      summary: Get quota usage for current tenant
      parameters:
        - name: quotaType
          in: query
          schema:
            type: string
          description: Optional filter by quota type
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuotaUsageDto'
  /grc/subscriptions/check-quota:
    post:
      tags: [Subscriptions]
      operationId: checkQuota
      summary: Check if quota allows operation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quotaType
                - requestedAmount
              properties:
                quotaType:
                  type: string
                requestedAmount:
                  type: number
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  allowed:
                    type: boolean
                  currentUsage:
                    type: number
                  limit:
                    type: number
                  remaining:
                    type: number

# ============================================================
# COMPONENTS
# ============================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /connect/authorize
          tokenUrl: /connect/token
          scopes:
            grc:read: Read GRC data
            grc:write: Write GRC data
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    LocalizedString:
      type: object
      properties:
        en:
          type: string
        ar:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: string
...
    ProductDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          $ref: '#/components/schemas/LocalizedString'
        description:
          $ref: '#/components/schemas/LocalizedString'
        category:
          type: string
        isActive:
          type: boolean
        displayOrder:
          type: integer
        iconUrl:
          type: string
...
    TenantSubscriptionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        product:
          $ref: '#/components/schemas/ProductDto'
        pricingPlanId:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
        trialEndDate:
          type: string
          format: date
          nullable: true
        autoRenew:
          type: boolean
        isActive:
          type: boolean
        isExpired:
          type: boolean
    SubscriptionDetailDto:
      allOf:
        - $ref: '#/components/schemas/TenantSubscriptionDto'
        - type: object
          properties:
            cancelledAt:
              type: string
              format: date-time
              nullable: true
            cancellationReason:
              type: string
    QuotaUsageDto:
      type: object
      properties:
        quotaType:
          type: string
        currentUsage:
          type: number
        limit:
          type: number
          nullable: true
        unit:
          type: string
        remaining:
          type: number
          nullable: true
        resetDate:
          type: string
          format: date
          nullable: true
        percentageUsed:
          type: number
          nullable: true
        isUnlimited:
          type: boolean
    SubscribeInput:
      type: object
      required:
        - productId
        - pricingPlanId
      properties:
        productId:
          type: string
          format: uuid
        pricingPlanId:
          type: string
          format: uuid
    CancelSubscriptionInput:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          minLength: 10
    UpgradeSubscriptionInput:
      type: object
      required:
        - productId
        - pricingPlanId
      properties:
        productId:
          type: string
          format: uuid
        pricingPlanId:
          type: string
          format: uuid